Proper type (most uint64_t for watershed changes)

+++ atomic_chunk.cpp
@@ -62,14 +62,14 @@ void processData(const AffinityExtractor<Ts, Ta, ConstChunkRef<Ta,4> > & affinit
     std::ofstream incomplete(str(boost::format("incomplete_edges_%1%.dat") % path), std::ios_base::binary);
     std::ofstream complete(str(boost::format("complete_edges_%1%.dat") % path), std::ios_base::binary);
 
-    std::vector<std::pair<float,int> > me;
+    std::vector<std::pair<float,uint64_t> > me;
     for (const auto & kv : edges) {
         const auto k = kv.first;
         if (incomplete_segments.count(k.first) > 0 && incomplete_segments.count(k.second) > 0) {
             incomplete_segpairs.push_back(k);
         } else {
             complete_segpairs.push_back(k);
-            me.push_back(meanAffinity<float, int>(edges.at(k)));
+            me.push_back(meanAffinity<float, uint64_t>(edges.at(k)));
         }
     }
     std::vector<int64_t> areas(complete_segpairs.size(),0);
@@ -78,7 +78,7 @@ void processData(const AffinityExtractor<Ts, Ta, ConstChunkRef<Ta,4> > & affinit
     std::iota(idx.begin(), idx.end(), 0);
 
     auto f_rlme = QtConcurrent::map(idx, [&](auto i)
-            {auto s = reweightedLocalMeanAffinity<float,int>(edges.at(complete_segpairs.at(i)));
+            {auto s = reweightedLocalMeanAffinity<float,uint64_t>(edges.at(complete_segpairs.at(i)));
              affinities[i] = s.first;
              areas[i] = s.second;
              });
@@ -90,7 +90,7 @@ void processData(const AffinityExtractor<Ts, Ta, ConstChunkRef<Ta,4> > & affinit
 
     f_rlme.waitForFinished();
 
-    for (int i = 0; i != complete_segpairs.size(); i++) {
+    for (size_t i = 0; i != complete_segpairs.size(); i++) {
         const auto & p = complete_segpairs[i];
         complete << std::setprecision (17) << p.first << " " << p.second << " " << me[i].first << " " << me[i].second << " ";
         complete << std::setprecision (17) << p.first << " " << p.second << " " << affinities[i]<< " " << areas[i] << std::endl;
@@ -114,9 +114,9 @@ int main(int argc, char * argv[])
 
     //FIXME: wrap these in classes
     bio::mapped_file_source seg_file;
-    size_t seg_bytes = sizeof(int32_t)*dim[0]*dim[1]*dim[2];
+    size_t seg_bytes = sizeof(uint64_t)*dim[0]*dim[1]*dim[2];
     seg_file.open("seg.raw", seg_bytes);
-    ConstChunkRef<int32_t,3> seg_chunk (reinterpret_cast<const int32_t*>(seg_file.data()), boost::extents[Range(offset[0], offset[0]+dim[0])][Range(offset[1], offset[1]+dim[1])][Range(offset[2], offset[2]+dim[2])], boost::fortran_storage_order());
+    ConstChunkRef<uint64_t,3> seg_chunk (reinterpret_cast<const uint64_t*>(seg_file.data()), boost::extents[Range(offset[0], offset[0]+dim[0])][Range(offset[1], offset[1]+dim[1])][Range(offset[2], offset[2]+dim[2])], boost::fortran_storage_order());
     std::cout << "mmap seg data" << std::endl;
 
     bio::mapped_file_source aff_file;
@@ -125,7 +125,7 @@ int main(int argc, char * argv[])
     ConstChunkRef<float,4> aff_chunk (reinterpret_cast<const float*>(aff_file.data()), boost::extents[Range(offset[0], offset[0]+dim[0])][Range(offset[1], offset[1]+dim[1])][Range(offset[2], offset[2]+dim[2])][3], boost::fortran_storage_order());
     std::cout << "mmap aff data" << std::endl;
 
-    AffinityExtractor<int32_t, float, ConstChunkRef<float,4> > affinity_extractor(aff_chunk);
+    AffinityExtractor<uint64_t, float, ConstChunkRef<float,4> > affinity_extractor(aff_chunk);
     traverseSegments(seg_chunk, true, affinity_extractor);
     processData(affinity_extractor, output_path);
     std::cout << "finished" << std::endl;
